RegisterLessonInputFormViewFactory를 프로토콜로 변경해 추상화 함
